fix warnings

+++ sliceable_switch/fdb.c
@@ -129,10 +129,10 @@ update_fdb( hash_table *fdb, const uint8_t mac[ OFP_ETH_ALEN ], uint64_t dpid, u
       return true;
     }
 
-    warn( "Failed to update fdb because host move detected in %d sec "
+    warn( "Failed to update fdb because host move detected in %" PRId64 " sec "
           "( mac = %02x:%02x:%02x:%02x:%02x:%02x, "
           "dpid = %#" PRIx64 " -> %#" PRIx64 ", port = %u -> %u ).",
-          HOST_MOVE_GUARD_SEC,
+          ( int64_t ) HOST_MOVE_GUARD_SEC,
           mac[ 0 ], mac[ 1 ], mac[ 2 ], mac[ 3 ], mac[ 4 ], mac[ 5 ],
           entry->dpid, dpid, entry->port, port );
